Stub in new spec_runner_helper
to run with the wrapper.  Also update Travis CI to install the
distribution for use in Cucumber scenarios
