+++ lambdas-and-strings/features/support/hooks.rb
@@ -1,3 +1,9 @@
+require_relative './helpers/runners/distribution_runner'
+
+Before('@distribution') do
+  set_spec_runner_helper DistributionRunner.new
+end
+
 Around('@log_commands') do |_scenario, block|
   original_logger = logger
